frost-core: split part of lib.rs into traits.rs and serialization.rs
